Remove async/await non-benefits from the benefits list (#10)
These aren't benefits -- they're rationalizations that the *negatives* of using async/await may be reduced in the future. Returning a bare `Task` makes any performance and compiler improvements in the future irrelevant because you've already optimized the time to 0.

I'm not arguing against the entire section -- I just think these two points are detracting and should be removed from the benefits list.